backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
PyMuPDF==1.24.0
langchain-google-genai==1.0.6
langchain-text-splitters==0.2.0
cloudinary==1.40.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pgvector==0.2.5
Pillow==10.3.0
httpx==0.27.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-pptx==0.6.23
python-docx==1.1.0
bcrypt==4.1.3
google-generativeai==0.5.4
